Extract timeline builder functions to reduce main.go surface.
Group timeline assembly (build + decorate) in timeline_builder.go so stream-timeline refactor can proceed in smaller, reviewable PRs.
